backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
aiosqlite==0.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
ortools==9.15.6755
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.1
pytest-asyncio==0.23.8
